Build fixed
git-svn-id: svn+ssh://svn.yandex.ru/yt/trunk@950 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ yt/ytlib/chunk_holder/block_store.cpp
@@ -186,7 +186,7 @@ TCachedBlock::TPtr TBlockStore::PutBlock(const TBlockId& blockId, const TSharedR
 {
     LOG_DEBUG("Putting block into store (BlockId: %s, BlockSize: %d)",
         ~blockId.ToString(),
-        data.Size());
+        static_cast<int>(data.Size()));
 
     return BlockCache->Put(blockId, data);
 }
